requirements.txt
APScheduler>=3.8.1
argon2-cffi>=21.3.0
dicebear==0.2.15
email-validator==1.1.3
Flask>=2.0.2,<3.0.0
Flask_Limiter>=2.0.4
Jinja2>=3.0.3,<4.0.0
requests>=2.28.0
Pillow>=9.0.0
setuptools>=60.1.0
WTForms>=3.0.0,<4.0.0
phonenumbers==8.12.42
qrcode>=7.3.1
pyotp>=2.6.0
google-api-python-client>=2.51.0
google-auth-httplib2>=0.1.0
google-auth-oauthlib>=0.5.2
CacheControl>=0.12.11
mysql-connector-python>=8.0.29
google-cloud-kms>=2.11.2
crcmod>=1.7
cryptography>=37.0.2
Authlib>=1.0.1
# splunk-appinspect